(experimental) allow dialects to have different values of supportsLimit and supportsOrderBy
