+++ packages/malloy/src/dialect/standardsql/functions/standardsql_functions.ts
@@ -22,10 +22,12 @@
  */
 
 import {FUNCTIONS} from '../../functions';
+import {fnStringAgg} from './string_agg';
 import {fnChr} from './chr';
 import {fnPi} from './pi';
 
 export const STANDARDSQL_FUNCTIONS = FUNCTIONS.clone();
 STANDARDSQL_FUNCTIONS.add('pi', fnPi);
 STANDARDSQL_FUNCTIONS.add('chr', fnChr);
+STANDARDSQL_FUNCTIONS.add('string_agg', fnStringAgg);
 STANDARDSQL_FUNCTIONS.seal();
